requirements.txt
requests==2.28.1
djangorestframework==3.14.0